정책 추천 탭 버그 수정
