+++ src/pages/recommend/policyRecommendTab.vue
@@ -3,17 +3,14 @@
   <div>
     <div v-if="goalRecommendations.length">
       <div v-for="goal in goalRecommendations" :key="goal.goalId" class="mb-4">
-        <!-- 목표 정보 -->
-        <h5 class="goal-title">
-          🎯 목표: {{ goal.keyword }} (ID: {{ goal.goalId }})
-        </h5>
-        <hr />
-
-        <!-- 추천 정책 카드 리스트 (가로 스크롤) -->
-        <div v-if="goal.recommendations.length">
-          <RecommendCardList :cards="goal.recommendations" type="policy" />
+        <div class="goal-header">
+          <h5 class="goal-title">{{ goal.goalName }}</h5>
+          <span class="goal-keyword">
+            #{{ keywordMap[goal.keyword] ?? goal.keyword }}
+          </span>
         </div>
-        <p v-else class="text-muted">추천할 정책이 없습니다.</p>
+        <RecommendCarousel :cards="goal.recommendations" />
+        <hr />
       </div>
     </div>
     <div v-else class="text-muted text-center">목표가 없습니다.</div>
@@ -23,13 +20,22 @@
 <script setup>
 import { onMounted, ref } from 'vue';
 import recommendPolicyAPI from '@/api/recommendPolicyApi';
-import RecommendCardList from './recommendCardList.vue';
-
+import RecommendCarousel from './policy/recommendCarousel.vue';
 const goalRecommendations = ref([]);
+
+const keywordMap = {
+  MARRIAGE: '결혼',
+  EMPLOYMENT: '취업',
+  HOME_PURCHASE: '내집마련',
+  TRAVEL: '여행',
+  EDUCATION_FUND: '학자금',
+  HOBBY: '취미',
+};
 
 onMounted(async () => {
   try {
     const res = await recommendPolicyAPI.getAllRecommendedPolicies();
+    console.log('✅ 받은 추천 데이터:', res);
     goalRecommendations.value = res ?? [];
   } catch (e) {
     console.error('❌ 전체 추천 로딩 실패:', e);
@@ -37,3 +43,32 @@ onMounted(async () => {
   }
 });
 </script>
+
+<style>
+.goal-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 0 5px 12px;
+}
+
+.goal-title {
+  font-size: 16px;
+  font-weight: 600;
+  margin: 0;
+  display: flex;
+  align-items: center;
+}
+
+.goal-title .icon-left {
+  margin-right: 6px;
+  font-size: 14px;
+  color: #555;
+}
+
+.goal-keyword {
+  font-size: 13px;
+  color: #888;
+  font-weight: 500;
+}
+</style>
